# chitchathub/src/main/java/com/sid/chitchathub/messages/MessageResponse.java
package com.sid.chitchathub.messages;

import lombok.*;

import java.time.LocalDateTime;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
@Builder
public class MessageResponse {

    private Long id;
    private String content;
    private MessageType type;
    private MessageState state;
    private LocalDateTime createdAt;
    private byte[] media;
    private String senderId;
    private String receiverId;
}